backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
pydantic==2.5.3

# TensorFlow - uncomment when adding real model
tensorflow==2.15.0
